Adequação dos models e admins
