+++ core/apps.py
@@ -3,4 +3,4 @@ from django.apps import AppConfig
 
 class CoreConfig(AppConfig):
     name = 'core'
-    verbose_name='Núcleo'
+    verbose_name='Admin do Assinaki'
